rename(dto): 패키지도 관심사/역할을 표현한다
