+++ src/main/java/blackjack/view/dto/DealerInitialHand.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.view.dto;
 
 import blackjack.domain.participant.Dealer;
 
@@ -6,7 +6,7 @@ public record DealerInitialHand(
         String nickname,
         String cardDisplayName
 ) {
-
+    
     public static DealerInitialHand from(Dealer dealer) {
         return new DealerInitialHand(
                 dealer.getNickname(),
