+++ src/main/java/blackjack/view/dto/ParticipantHandScore.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.view.dto;
 
 import blackjack.domain.participant.Participant;
 import java.util.List;
@@ -8,7 +8,7 @@ public record ParticipantHandScore(
         List<String> cardDisplayNames,
         int totalScore
 ) {
-
+    
     public static ParticipantHandScore from(Participant participant) {
         return new ParticipantHandScore(
                 participant.getNickname(),
@@ -16,13 +16,13 @@ public record ParticipantHandScore(
                 participant.getScore()
         );
     }
-
+    
     public static List<ParticipantHandScore> listOf(List<Participant> participants) {
         return participants.stream()
                 .map(ParticipantHandScore::from)
                 .toList();
     }
-
+    
     private static List<String> toCardDisplayNames(Participant participant) {
         return participant.getCards().stream()
                 .map(CardDisplayName::from)
