+++ src/main/java/blackjack/view/dto/PlayerNames.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.view.dto;
 
 import blackjack.domain.participant.Player;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.List;
 public record PlayerNames(
         List<String> playerNames
 ) {
-
+    
     public static PlayerNames from(List<Player> players) {
         return new PlayerNames(
                 players.stream()
