+++ src/main/java/blackjack/view/dto/TotalWinningResult.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.view.dto;
 
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.GameResult;
@@ -11,16 +11,16 @@ public record TotalWinningResult(
         long dealerLossCount,
         List<PlayerGameResult> playerResults
 ) {
-
+    
     public static TotalWinningResult of(Dealer dealer, List<Player> players) {
         Map<GameResult, Long> dealerResult = dealer.determineGameResult(players);
         long dealerWinCount = dealerResult.getOrDefault(GameResult.WIN, 0L);
         long dealerLoseCount = dealerResult.getOrDefault(GameResult.LOSE, 0L);
-
+        
         List<PlayerGameResult> playerResults = players.stream()
                 .map(player -> PlayerGameResult.of(player, dealer))
                 .toList();
-
+        
         return new TotalWinningResult(dealerWinCount, dealerLoseCount, playerResults);
     }
 }
